[refactor] JsonNode를 Command로 바로 변경
