--- src/main/java/team7/inplace/crawling/client/dto/RawVideoInfo.java
@@ -1,8 +1,0 @@
-package team7.inplace.crawling.client.dto;
-
-public record RawVideoInfo(
-        String videoId,
-        String videoTitle,
-        String address
-) {
-}
